backend/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
langchain-nvidia-ai-endpoints
python-dotenv
requests
beautifulsoup4
